pyproject.toml
[tool.poetry]
name = "newspaper"
version = "0.1.0"
description = "The most truthful newspaper"
authors = ["Sergey Dyachok <sergey@sdyachok.com.ua>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
psycopg2-binary = "^2.8.5"
Pillow = "^7.2.0"
django-imagekit = "^4.0.2"
django-bootstrap4 = "^2.2.0"
django-extensions = "^3.0.2"
django-summernote = "^0.8.11"
django-gravatar2 = "^1.4.4"
django-haystack = {version = "3.0b2"}
Whoosh = "^2.7.4"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
black = {version = "19.10b0", allow-prereleases = true}
flake8 = "^3.8.3"
django-debug-toolbar = "^2.2"
Werkzeug = "^1.0.1"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.eggs
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | docs
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
